OLMIS-1044 Add centralized logging configuration
Add logging configuration using a central logback file to each microservice. A logbook file is provided in blue as a template to show how different parts of the system can be logged differently. This logback file is mounted into each microservice and is used by the Spring Boot jar file when starting through the -Dlogging.config parameter.