change guidance from iterations to hold-for
We're now at a place where using hold-for makes more sense than using iterations.  Iterations is problematic for the test writer to judge how many is a useful sample size, especially for things like p90.  We've also cleared away some tech debt that means using hold-for for short periods of times makes a lot of sense for performance testing as well as stress testing.