OLMIS-3316, move openlmis/config/.env into this repo and place service versions in .env
The config/.env file has been move to be a settings-sample.env file which you're supposed to copy to settings.env and modify with your values.  This file is still passed through to every service to help configure them.  A new .env file is located in this repo and is meant to be the .env file of the docker-compose.yml.  It is meant to hold information such as which version of the service should I run, or which host port should I map a specific service's port to.  It is not passed through to each service.  Seperating these environment/config files in this way creates a nice division between configuring the ref distro, and configuring the services the ref distro is going to run.