Initializing NiFi (#26)
* OLMIS-5095: Download NiFi Toolkit into Container

Add a script to the NiFi container for downloading the NiFi toolkit
binary, if it doesn't already exist. The NiFi Toolkit contains the NiFi
CLI tool.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* OLMIS-4626: Add Script For Initializing NiFi

NiFi needs to be initialized after the NiFi service fully starts.
Initialization will involve:
- Creating connections to NiFi Registries
- Importing Flows from those Registries

The nifi/scripts/registry.sh script has been added to do that.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>

* Harden Logic For Extracting NiFi Names

The logic in preload.sh for extracting the names of NiFi Registry
clients and their IDs would likely break if a registry client name was
not one word. This commit fixes that.

Co-authored-by: Morris Mukiri <morrismukiri@gmail.com>
Co-authored-by: Jason Rogena <jasonrogena@gmail.com>
Signed-off-by: Jason Rogena <jasonrogena@gmail.com>