Change base image of openlmis/awscli to alpine
Since openlmis/dev is based on an image based on alpine,
we can reduce the image size by using alpine directly.

By using the same version of alpine as openlmis/dev base java  image,
we can keep the benefit of having to download less layers when used  in
the same environment as openlmis/dev.

Signed-off-by: Morris Mukiri <morrismukiri@gmail.com>