OLMIS-6183, simplify proxy options
Focus is on the nginx config and changing the proxy pass options.  Primarly forwarding the Host header that will include the port as well as the requested address (e.g. nifi.local) and the port.  This allows Nginx to be bound to ports other than 80 and 443.  Further reduced the expected use-cases be assuming that Nginx as a reverse proxy will fully encapsulate Nifi and Superset, and that if a user wants to use SSL, either the termination will be done at Nginx and the configuration will include certs, or the termination may happen a level above Nginx (e.g. an AWS ELB) in which case running a server on port 443 is unneeded.