OLMIS-6858 Enable streaming architecture for reporting stack
This is to start the data pump feature for the reporting stack. The components of the streaming architecture are added (Kafka, Connect, etc.) and the first set of connectors (for reference data) are registered. There is one source connector for all reference data tables and one sink connector for most reference data tables. The additional sink connectors are for tables that are not identified by an id field, but rather by a combination of ids.